minor refactorings, e.g. removing unnecessary namespaces, renamed scipt file, etc.
