+++ lib/paasal/scripts/initialize_config_defaults.rb
