--- lib/paasal_api/api/errors/bad_request_error.rb
@@ -1,12 +1,0 @@
-module Paasal
-  module API
-    module Errors
-      class BadRequestError < ApiError
-        # initialize with default error to be 400, bad request
-        def initialize(message, ui_error = ErrorMessages::BAD_REQUEST)
-          super(message, ui_error)
-        end
-      end
-    end
-  end
-end
